Move the quickbooks stuff to the admin page. Fix logging, currently set to debug
