+++ metabrainz/templates/admin/master.html
@@ -15,5 +15,6 @@
     <a href="/">MetaBrainz</a>
     ::
     <a href="/admin/">BDFL Zone</a>
+    <a href="/admin/quickbooks">QuickBooks</a>
   </span>
 {% endblock %}
